EX3 partial 1
'Cast Rays' excercise partly done. Added Layer 8 label 'ISM collider' and ISMReverb script efited. There are some remaining issues: when the final ray is traced to the source, it doesn't notice obstruction and when the image source is beyond a parallel surface (e.g. the other side of the cube), the ray is not cast.